Save system revamp
Had to unfortunately revamp the save system because it wasn't working for combat encounters. The fact that I had to flip levers in between battles meant I had to save the levers' states, and that meant I really neeeded to revamp saves to work more flexibly. We're halfway there. Tomorrow, more of that and then finally working on the gameplay loop and its elements to mess around with.